Fix high-frequency compiler warnings: Batch 7 - sign-compare and uniform_int_distribution<size_t> issues
- spatial_index.cpp:1296: Change loop counter from int to size_t for proper size_t comparison
- product_quantizer.cpp:533,553: Replace std::uniform_int_distribution<size_t> with uint64_t
- ann_index.cpp:110: Replace std::uniform_int_distribution<size_t> with uint64_t
- cross_shard_transaction.cpp:2512: Replace uniform_int_distribution<size_t> with uint64_t
- rotate_completion.cpp:254,273: Replace uniform_int_distribution<size_t> with uint64_t
- anomaly_detection.cpp:300,307,662,667: Replace uniform_int_distribution<size_t> with uint64_t (2 locations)
- voice_liveness_detector.cpp:297: Replace uniform_int_distribution<size_t> with uint64_t
- rabitq.cpp:302,306,375: Replace uniform_int_distribution<size_t> with uint64_t
- tensor_router.cpp:142,144: Replace uniform_int_distribution<std::size_t> with uint64_t

Total changes: 9 files, ~21 lines modified
Impact: Eliminates -Wsign-compare and -Wunused-variable warnings across index and related modules

+++ src/analytics/anomaly_detection.cpp
@@ -297,14 +297,14 @@ ITree buildITree(const FeatureMatrix &fm, const std::vector<size_t> &indices, in
             continue;
         }
 
-        std::uniform_int_distribution<size_t> feat_dist(0, n_features - 1);
+        std::uniform_int_distribution<uint64_t> feat_dist(0, n_features - 1);
         size_t feat              = 0;
         double fmin              = 0.0;
         double fmax              = 0.0;
         bool found_split_feature = false;
 
         for (size_t attempt = 0; attempt < n_features; ++attempt) {
-            feat = feat_dist(rng);
+            feat = static_cast<size_t>(feat_dist(rng));
             fmin = std::numeric_limits<double>::max();
             fmax = std::numeric_limits<double>::lowest();
             for (size_t i : idx) {
@@ -659,12 +659,12 @@ struct AnomalyDetector::Impl {
             std::mt19937 rng(42);
             forest.clear();
             forest.reserve(static_cast<size_t>(cfg.n_estimators));
-            std::uniform_int_distribution<size_t> idx_dist(0, data.size() - 1);
+            std::uniform_int_distribution<uint64_t> idx_dist(0, data.size() - 1);
 
             for (int t = 0; t < cfg.n_estimators; ++t) {
                 std::vector<size_t> sample(static_cast<size_t>(sub_size));
                 for (auto &s : sample) {
-                    s = idx_dist(rng);
+                    s = static_cast<size_t>(idx_dist(rng));
                 }
                 forest.push_back(buildITree(fm, sample, 0, hl, rng));
             }
